feat Improved the widget's default error messages customization.
The implementation code is now cleaner and more documented, as well as for the test. A little bit of redundancy checking has been introduced in the tests just to make sure that if the order of some methods changes or if some are removed or whatever the tests will still work and be comprehensive.

Also fixed a small mistake in AutoComplete's default error messages test.

An unwanted `return` statement was remaining at the end of the `$constructor` of the test. It was not breaking anything regarding our class system implementation, but this remains incorrect.