fix - Fixed the escape behavior for several widgets
That way widgets such as the dialog can be closed with this key.

Now their associated controllers properly stop the behavior of the events as needed.

I had to add an additional property to the controllers' report model so that they can tell to prevent the default behavior of the browsers without making the propagation stop. This is needed to keep a bug fix while enabling the new behavior on escape.