fix On Chrome, caret was put at the beginning of inputs upon up arrow keydown event
Putting the caret at the beginning of inputs when the up arrow is pressed is a native behavior of Chrome.

However, since we manage dropdown-based widgets in a way where the focus is kept inside the input, while what we actually manipulate with the up and down arraows is inside the dropdown, this behavior should not happen. Especially since it troubles the typing experience (if the user wants to edit the text directly while the dropdown is still open).

That's why for all dropdown based widget, if the popup is open, the default behavior of the up arrow gets discarded. This has been made for Chrome only.